pyproject.toml
[project]
name = "ml2code"
version = "0.0.1"
description = "ML2Code"
readme = "README.md"
authors = [
    { name = "Jared Hulbert", email = "jaredeh@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.1",
    "onnx>=1.18.0",
    "onnx2torch>=1.5.15",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.1",
    "pyaml>=25.5.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]

[project.scripts]
ml2code = "ml2code:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ml2code"]
